requirements.txt
Pillow
numpy
tensorflow
opencv-python
fastapi
uvicorn
python-multipart
scikit-learn
joblib
